requirements.txt
Pillow==10.0.0
PySimpleGUI==4.60.5
